Correct ListsReducer tests
